requirements.txt
markdown==3.4.3
python-docx==0.8.11 